pyproject.toml
[tool.poetry]
name = "creator-report-unifor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^1.1.2"
pandas = "^2.2.3"
streamlit = "^1.39.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
